+++ scripts/update_poetry.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-SCRIPT=$(realpath "$0")
-SCRIPTPATH=$(dirname "$SCRIPT")
-APP_ROOT="$(dirname "$SCRIPTPATH")"
-cat $APP_ROOT/requirements.txt | xargs poetry add -C $APP_ROOT/envs/